# Improve ChildSelector performance