--- a/src/AngleSharp/Css/Dom/Internal/FirstChildSelector.cs
+++ b/src/AngleSharp/Css/Dom/Internal/FirstChildSelector.cs
@@ -21,20 +21,38 @@ namespace AngleSharp.Css.Dom
 
             if (parent != null)
             {
-                var n = Math.Sign(Step);
-                var k = 0;
-
-                for (var i = 0; i < parent.ChildNodes.Length; i++)
+                // remove interface dispatch overhead
+                if (parent.ChildNodes is NodeList nodeList)
                 {
-                    if (parent.ChildNodes[i] is IElement child && Kind.Match(child, scope))
-                    {
-                        k += 1;
+                    return DoMatch(new ConcreteNodeListAccessor(nodeList), element, scope);
+                }
 
-                        if (child == element)
-                        {
-                            var diff = k - Offset;
-                            return diff == 0 || (Math.Sign(diff) == n && diff % Step == 0);
-                        }
+                return DoMatch(new InterfaceNodeListAccessor(parent.ChildNodes), element, scope);
+            }
+
+            return false;
+        }
+
+        private Boolean DoMatch<T>(T nodes, IElement element, IElement? scope) where T : INodeListAccessor
+        {
+            var step = Step;
+            var n = Math.Sign(step);
+            var k = 0;
+            var kind = Kind;
+            var matchAll = ReferenceEquals(Kind, AllSelector.Instance);
+            var offset = Offset;
+            var length = nodes.Length;
+
+            for (var i = 0; i < length; i++)
+            {
+                if (nodes[i] is IElement child && (matchAll || kind.Match(child, scope)))
+                {
+                    k += 1;
+
+                    if (child == element)
+                    {
+                        var diff = k - offset;
+                        return diff == 0 || (Math.Sign(diff) == n && diff % step == 0);
                     }
                 }
             }