--- a/src/AngleSharp/Dom/Internal/NodeList.cs
+++ b/src/AngleSharp/Dom/Internal/NodeList.cs
@@ -13,7 +13,7 @@ namespace AngleSharp.Dom
     {
         #region Fields
 
-        private readonly List<Node> _entries;
+        internal readonly List<Node> _entries;
 
         /// <summary>
         /// Gets an empty node-list. Shouldn't be modified.
@@ -96,5 +96,58 @@ namespace AngleSharp.Dom
         IEnumerator IEnumerable.GetEnumerator() => _entries.GetEnumerator();
 
         #endregion
+    }
+
+    /// <summary>
+    /// Helper interface which can remove interface dispatch overhead when used in combination with generic methods.
+    /// </summary>
+    internal interface INodeListAccessor
+    {
+        Int32 Length { get; }
+        INode this[Int32 index] { get; }
+    }
+
+    internal readonly struct ConcreteNodeListAccessor : INodeListAccessor
+    {
+        private readonly List<Node> _nodeList;
+
+        public ConcreteNodeListAccessor(NodeList nodeList)
+        {
+            _nodeList = nodeList._entries;
+        }
+
+        public Int32 Length
+        {
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            get => _nodeList.Count;
+        }
+
+        public INode this[Int32 index]
+        {
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            get => _nodeList[index];
+        }
+    }
+
+    internal readonly struct InterfaceNodeListAccessor : INodeListAccessor
+    {
+        private readonly INodeList _nodeList;
+
+        public InterfaceNodeListAccessor(INodeList nodeList)
+        {
+            _nodeList = nodeList;
+        }
+
+        public Int32 Length
+        {
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            get => _nodeList.Length;
+        }
+
+        public INode this[Int32 index]
+        {
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            get => _nodeList[index];
+        }
     }
 }